requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
autopep8==1.5.2
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
distlib==0.3.0
Django==3.0.5
django-knox-rest==4.1.0
django-sslserver==0.22
django-storages==1.9.1
djangorestframework==3.11.0
filelock==3.0.12
google-api-core==1.17.0
google-auth==1.15.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
protobuf==3.12.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pylint==2.5.0
PyMySQL==0.9.3
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.20
wrapt==1.12.1
zipp==3.1.0
